batch controller added, able to add products and corresponding events
user id is int, citizenId is long. böyle yazılmayan citizenId varsa dikkat

+++ src/main/java/com/agritrust/service/CertificateReadable.java
@@ -7,5 +7,5 @@ import com.agritrust.entity.CertificateEntity;
 public interface CertificateReadable extends Readable<CertificateEntity, Long>{
 	CertificateEntity getByCertificateNumber(String certificateNumber);
 	List<CertificateEntity> getAllValid();
-	List<CertificateEntity> getByAuditor(Long auditorId);
+	List<CertificateEntity> getByAuditor(Integer auditorId);
 }
